requirements.txt
asgiref==3.11.0
Django==5.2
django-cors-headers==4.9.0
django-filter==25.2
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.0
PyJWT==2.11.0
sqlparse==0.5.5
tzdata==2025.3
gunicorn
